% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pipelines_HDF5_script.R
\name{run_all_pipelines}
\alias{run_all_pipelines}
\title{Run all Pipelines}
\usage{
run_all_pipelines(HDF5_file = NULL, verbose = FALSE)
}
\arguments{
\item{verbose}{Logical; if TRUE, prints additional messages during processing.}

\item{mat}{A matrix or vector containing quality metrics.}

\item{df}{Data frame having columns \code{Chr}, \code{Pos}, \code{Ref}, and \code{Alt}.}

\item{expected_variants}{The expected variant numbers}
}
\value{
A list with elements:
\describe{
\item{common_variants}{Charactor vector of variant IDs commonon to both pipelines.}
\item{optima_unique}{charactor vector of variant IDs unique to the Optima pipelin.}
\item{tapestri_unique}{character vector of variant IDs unique to the Tapestri pipeline.}
}

A numeric vector of summary values, one per variant.

Data frame in BED format with columns: \code{Chr}, \code{start_coord}, \code{end_coord}, and \code{extra_info}

BED-format data frame of passed variants

BED-format data frame of filtered variants
}
\description{
A given matrix or vector of quality matrics, the helper function returns a summary
vector of length equal to expected number of variants. If the matrix has dimension(cell x variants),
it uses colMeans; if transposed(variants x cells), then rowMeans.

Converts data frame having variant information(columns: chr, pos, ref, alt) into
BED-format. The start & end coordinates are calculated and constructs an extr_info field.
}
\examples{
# For a matrix of dimensions 4858 x 82247:
# summarised <- summarise_matrix(my_matrix, 82247)

}
